Spec v1.0

+++ include/host/Enclave.hpp
@@ -68,7 +68,7 @@ class Enclave {
       const char* eapppath, const char* runtimepath, Params _params,
       uintptr_t alternatePhysAddr);
   Error destroy();
-  Error run(uintptr_t* ret = nullptr);
+  Error run(unsigned long *ret = nullptr);
 };
 
 uint64_t
